+++ Synchronization.Service/ISyncServiceClient.cs
@@ -6,15 +6,15 @@ using System.Threading;
 using System.Threading.Tasks;
 
 namespace Synchronization.Service {
-    public interface IDitherServiceClient {
+    public interface ISyncServiceClient {
         void RegisterSync();
         void UnregisterSync();
         Task Register();
         Task Unregister();
-        Task<bool> WaitForSync(CancellationToken ct, TimeSpan timeout);
-        Task WaitForDither(CancellationToken ct, TimeSpan timeout);
-        Task SetDitherInProgress(CancellationToken ct);
-        Task SetDitherCompleted(CancellationToken ct);
+        Task<bool> WaitForSyncStart(CancellationToken ct, TimeSpan timeout);
+        Task WaitForSyncComplete(CancellationToken ct, TimeSpan timeout);
+        Task SetSyncInProgress(CancellationToken ct);
+        Task SetSyncComplete(CancellationToken ct);
         Task AnnounceToSync(bool canLead, CancellationToken ct);
         Task<string> Ping(CancellationToken ct);
     }